+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - VIMEET-2284 - Delete question webinar
 - VIMEET-2184 - Speakers can prepare before their webinar
 - VIMEET-2234 - API endpoint to access list of happenings
+- VIMEET-2278 - Participant's program displayed according his timezone
 - VIMEET-2253 - Increment unread message counters synchronously
 - Add health check route
 - VIMEET-2275 - Close notification private chat